requirements.txt
# Runtime dependencies
schedule>=1.2.0
discord.py>=2.3.0
python-dotenv>=1.0.0
